requirements.txt
accelerate==1.10.1
basicsr==1.4.2
codeformer-pip==0.0.4
diffusers==0.35.1
facexlib==0.3.0
gdown==5.2.0
gradio==5.44.1
gradio_client==1.12.1
huggingface-hub==0.34.4
numpy==2.2.6
opencv-python==4.11.0.86
pillow==11.3.0
requests==2.32.5
scikit-image==0.25.2
scipy==1.16.1
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
transformers==4.56.1
triton==3.4.0
uvicorn==0.35.0